threat_detection/requirements.txt
Django==4.2.7
tensorflow-cpu==2.13.0  # Use CPU version for smaller image
librosa==0.10.1
numpy==1.24.4
gunicorn==21.2.0
psycopg2-binary==2.9.7  # For PostgreSQL support
python-dotenv==1.0.0
sounddevice==0.4.6
scipy==1.11.3
tensorflow-hub==0.14.0
channels==4.0.0
daphne==4.0.0
uvicorn==0.23.2
channels-redis==4.1.0  # Required for Redis channel layer